Docking the details panel to the right hand side
Prevent controls from appearing on the panel before a layer is added to the map.
Basically when the panel first rendered some controls that don't do anything are
on it, then when a layer is added they disappear and relevant ones are displayed.
This change prevents those initial unmapped controls from appearing.