#1098 Added calls to loading layer JS on animation creation
Layers were loading super fast for me so I couldn't see the animation but if I
block in the debugger they definitely appear so I am going to say this is done.
The calls need to be explicit rather than relying on the loadstart and loadend events
of the layer as they are both quite brief, it's the creation of the layer that seems
to take time so the calls to the spinner surround that.