#1140 Fixed "Can't select features in layer added from search page"
Querying for feature info required a layer to be queryable and the version to be specified.
It is not possible to determine a layer's queryable state from geonetwork search results so
that restriction has been removed.  In addition the layer version is derived from the server type,
the server type is now set on a search result layer by interrogating the protocol.