Fix for the assignment to transient property layer.layers
With a recent fix the layers.layer property is transient and should not be assigned to.
There was a code path when fetching the layers for a server that did this assignment after
filtering out black listed and inactive layers. This process should rightly be handled
dynamically returning a new list or use a presenter object to return the collection of
layers that can be displayed. The second option was chosen - and it feels good.
http://www.youtube.com/watch?v=7SPDfS4jt6E