Yet another fix for proxied responses
The GetFeatureInfo requests were failing as they were including a format=text/html parameter and the
ProxyController was using that content type on all GetFeatureInfo responses.  The GetFeatureInfo requests
no longer have a need for that parameter so it was removed.  This, once again, appears to have resolved the
issue.  I'd love to WTF is going on with the production apache setup that the fix applied elsewhere didn't just
work.