Fix for server request from menu
When clicking a folder icon that is a server (discovery) the app errors and nothing is returned. There was a
method call that was missed in a refactor - the method had been moved to the Server domain class but
the controller was still calling it as if it was local.