#1818 pretty up timestamps for display
There are some animated layer timestamps along the lines of 12:29:59 and 12:29:59.999 this fix rounds those
values to 12:30:00 for display purposes in the comboboxes.  The animation panel now uses actual date objects
rather than strings for the heavy lifting and these are formatted to strings for display.  Dates and times are displayed
in the browser (clients) timezone, this is the way JS works.