#1891 depth service was throwing a 500 error when no rows were returned
Depth service has been updated to return a 204 when no matching rows are found and portal now handles the
blank content accordingly.  Unfortunately grails/groovy changes the response code to 200 and I haven't found a
quick and easy way to pass the 204 through so the JS side of things handles a blank response as if there were no
response.