Fix menu caching
Previous changes resulted in the menu ceasing to be pulled from the cache before a menu was fetched from the db
and parsed into a MenuPresenter, however the optimisation to wrap the Menu in the presenter was performing fast
enough to have hidden that problem.  In addition once a menu had been edited and recached it was being cached
under an incorrect key resulting in the changes not appearing in the UI.  Will patch this to the current release.