Fixes #77 and #87
Removed references to Portal.app.config.hideLayerOptions because it isn't actually used.
Change when the right details panel is expanded on first load as this addresses the losing of
the expand button. There's a little bit of trickery where the right panel thinks it is expanded
when it isn't so you have to convince it is still collapsed.