Fix for #175
Doesn't feel like the tidiest solution, for some reason on the call to the feature info popup close
the Shadow.js code gets -1 as a value for the adjusts height in the popup which is an invalid value.
I couldn't trace where that value has come from as we certainly don't set it, catching the exception
appears to allow everything to work, am more than happy for someone to improve this solution!