Changes for #265
The fix: Only include the possible values for Strings and Dates as they're currently the only UI elements that use them.

While I was in there I also added sorting to the Filters so they will always appear in the same order in the Portal. The order is not currently customisable (it's just alphabetical by label).