Map drawing polygon type selection
Allows the user to select what type of bounding area they want to draw
on the map, options being a bounding box, polygon or none for navigating
the map. There are currently two instances of the control, one on the
AODAAC subset panel and one on the WFS subset panel, they cannot stay in
sync at the moment because the FilterGroupPanel is re-created on every
layer swap.