Update layer when time slider is clicked
The map would update if the animation slider was dragged but not if a
new position was set by clicking on the slider. This addresses that by
listening to the changecomplete event, there may be a more appropriate
event to listen to but I could not find it, I tried dragend for example.