getStyles always returns a map
Change the server-side API for ncWMS layers. We can save a bunch of checking code by changing getStyles() to always return a map. The client-side code will fail if the required information is missing, but the old code would have failed, too, and there is a catch block handling these failures (in JS).