Sort styles by palette first
In talking to GG about this issue he pointed out that the first thing to do when looking at a dataset is find an appropriate palette. When you have that you can try the arrow styles till you get one that suits. So he asked that the stlyes dropdown is sorted so that styles of the same palette are grouped together.