Align filter export script with extension XML and directory structure
Filters are stored in a file called filters.xml under
workspaces/$workspace/$data_store/$layer_name the data store name is not
retrievable at this stage so this script exports to
filters/$workspace/$layer_name/filters.xml.

Additionally export the XML in the format piped down as the OWS response
from the geoserver extension. Needs possible values to be removed as
they are not persisted to disk by the geoserver extension.