Feature toggle DB filters vs GeoServer filters
For the GeoServer filters we want to load the filters from the download layer if it is available. However, if we're using DB filters we need to get the filters from the map layer. So the feature toggle is used here to know which layer needs to be queried for filters.