Decide in client side which filter component to use
The main behaviour of this method is now converting GeoServer types --> UI Components.
As a temporary measure while the DB is still in play it will first convert
Portal types --> GeoServer types. This basically undoes some work that happens during
a WFS scan, but maintains compatability with the old behaviour.