DateFilter display updated
Uses 'label: value' format except that the label is not from the filter but
just a value from the i18n dictionary. This is so that the filter display on step 3
is consistent with step 2. Step 2 uses values from the dictionary, rather than the
actual filter labels, so that geographic bounds and date range are consitent across
all data collections.