Introduced library 'jquery.fileDownload' to do all the client-side download dirty work
The main upshot of this commit is that failure events now include the message returned by the server.
Note that a timeout is no longer applied after which a failure event will be raised, because the library doesn't allow this behaviour.  We *may* need to revisit this if it's a problem.