Changes from review comments
Using font awesome for spinner.  @pmbohm has put up his hand to change all spinners to be done this way :-)  so that's why I'm happy to go with that here.
I will put the console logging change up in a separate PR as I think it's worth some discussion within the team.