Fixes #1770
When coords were entered by hand a copy of the geometry was not being kept
for rolling-back after an error. This change fixes the problem but it's
not really clear why this change needed to go here so I will do some
renaming and rearranging after this to hopefully clarify things.