Reorder layers on accordion selection
The idea is that when an accordion is opened that layer is now selected and
we want to bring that layer to the top of the map (to ensure it is not obscured
by other layers). There may be a better way to move layers in the store, for
now I'm just removing and re-adding them.