Fixes related to #1835
Although #1835 doesn't appear to be there on master anymore, I noticed some other weird behaviour here.
Namely, that when a layer is added, removed, and the added again, the expected layer loading events don't fire, in turn meaning that the loading/error indicators on the Subset panel aren't disaplyed appropriately.

This fixes that, although it *may* affect tile loading performance, so we should keep an eye any issues related to that.