Fixes #1999
Don't attempt to load filters from the server for NcWms - currently, they are hard-coded in the client and trying to load from server just wipes out the desired filters with an empty list.  This in turn causes problems such as #1999.

Note that the proper fix is to treat NcWms collections like any other, i.e.:

* load filters from server
* model filters in exactly the same way as non-ncwms
* use the same UI components as non-ncwms