Fix PointTimeSeriesLatLonTest selenium test to match current functionality… (#2654)
* Fix PointTimeSeriesLatLonTest selenium test to match current functionality.
Evidently with the older depth service the latitude and longitude given in the FeatureInfo were whatever latitude and longitude were sent in the request. But with the new service the lat and lon represent the location of the depth data, which ends up being slightly different to the requested location.  So I've given the test a little leeway by checking if latitude and longitude are within 0.1 of their expected values. I also added a check to make sure a depth/elevation is actually being shown.

* Make difference absolute